Scene: increase the number of FPS samples, use fixed-point arithmetic
Using a higher number of samples (enough samples to account for the last
second or two of playback for e.g.) can be useful when comparing minor
changes in overall playback speed, where the behavior of multi-threaded
operations can make the value jitter with 8 samples (default).

Using fixed-point arithmetic means the average FPS can be updated
by subtracting the oldest FPS sample before adding the new value,
instead of having to average an array of floats every draw.
Increasing the number of samples now only uses a little more memory
(20kb at most).

The error margin from using fixed-point arithmetic is under 0.5
microseconds per frame - more than enough precision for FPS display.
A commented define is included that shows the error margin when enabled.

+++ source/blender/editors/include/ED_scene.hh
@@ -50,7 +50,7 @@ void ED_operatortypes_scene();
  * \param ltime: Time since the last update,
  * compatible with the result of #PIL_check_seconds_timer.
  */
-void ED_scene_fps_average_accumulate(struct Scene *scene, uchar fps_samples, double ltime)
+void ED_scene_fps_average_accumulate(struct Scene *scene, short fps_samples, double ltime)
     ATTR_NONNULL(1);
 /**
  * Calculate an average (if it's not already calculated).
